Remove from lock and cover
