+++ src/cards/cover-card/cover-card-config.ts
@@ -12,7 +12,6 @@ export type CoverCardConfig = LovelaceCardConfig &
     EntitySharedConfig &
     AppearanceSharedConfig &
     ActionsSharedConfig & {
-        display_battery?: boolean;
         show_buttons_control?: false;
         show_position_control?: false;
         show_tilt_position_control?: false;
@@ -22,7 +21,6 @@ export const coverCardConfigStruct = assign(
     lovelaceCardConfigStruct,
     assign(entitySharedConfigStruct, appearanceSharedConfigStruct, actionsSharedConfigStruct),
     object({
-        display_battery: optional(boolean()),
         show_buttons_control: optional(boolean()),
         show_position_control: optional(boolean()),
         show_tilt_position_control: optional(boolean()),
