+++ src/cards/lock-card/lock-card.ts
@@ -7,13 +7,13 @@ import {
     ActionHandlerEvent,
     computeRTL,
     handleAction,
-    hasAction, hasBattery,
+    hasAction,
     HomeAssistant,
     isAvailable,
     LockEntity,
     LovelaceCard,
     LovelaceCardEditor
-} from '../../ha';
+} from "../../ha";
 import "../../shared/badge-icon";
 import "../../shared/button";
 import "../../shared/card";
@@ -87,7 +87,6 @@ export class LockCard extends MushroomBaseCard implements LovelaceCard {
         const icon = this._config.icon || stateIcon(entity);
         const appearance = computeAppearance(this._config);
         const picture = computeEntityPicture(entity, appearance.icon_type);
-        const battery = Boolean(this._config?.display_battery) && hasBattery(entity);
 
         const rtl = computeRTL(this.hass);
 
@@ -104,7 +103,7 @@ export class LockCard extends MushroomBaseCard implements LovelaceCard {
                         })}
                     >
                         ${picture ? this.renderPicture(picture) : this.renderIcon(entity, icon)}
-                        ${battery ? this.renderBatteryBadge(entity) : this.renderBadge(entity)}
+                        ${this.renderBadge(entity)}
                         ${this.renderStateInfo(entity, appearance, name)};
                     </mushroom-state-item>
                     <div class="actions" ?rtl=${rtl}>
