+++ src/cards/vacuum-card/vacuum-card.ts
@@ -144,6 +144,22 @@ export class VacuumCard extends MushroomBaseCard implements LovelaceCard {
         `;
     }
 
+    protected renderBatteryBadge(entity: HassEntity): TemplateResult | null {
+        const level = entity.attributes.battery_level;
+        const color = level > 40 ? "var(--rgb-success)" : level > 20 ? "var(--rgb-warning)" : "var(--rgb-danger)";
+        const iconStyle = {
+            "--main-color": `rgb(${color})`,
+        };
+
+        return  html`
+              <mushroom-badge-icon
+                slot="badge"
+                style=${styleMap(iconStyle)}
+                icon=${entity.attributes.battery_icon}
+              ></mushroom-badge-icon>
+          `;
+    }
+
     static get styles(): CSSResultGroup {
         return [
             super.styles,
